[Feat] 03/04 SWLee2973 전체 레이아웃 및 피드 #16 - 전체 레이아웃 스크롤 방식 수정 - 비쥬얼 영역 디자인 업데이트 - 피드 및 그룹 헤더 조립
