+++ index.html
@@ -30,7 +30,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap" rel="stylesheet">    
   </head>
-  <body class="overflow-y-clip">
+  <body class="overflow-y-scroll">
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
