+++ package.json
@@ -20,6 +20,7 @@
     "react-helmet-async": "^2.0.4",
     "react-hook-form": "^7.50.1",
     "react-router-dom": "6.22.0",
+    "react-transition-group": "^4.4.5",
     "swiper": "^11.0.6",
     "tailwind-scrollbar-hide": "^1.1.7",
     "zustand": "^4.5.1"
