--- src/pages/GroupTest.jsx
@@ -1,5 +1,0 @@
-export const Component = () => {
-  return <div>모임임</div>;
-};
-
-Component.displayName = 'group';
