+++ src/pages/MyPage.jsx
@@ -1,11 +1,13 @@
-import { Outlet } from 'react-router-dom';
+import MainNavBar from '@/components/molecule/common/MainNavBar';
 
 export const Component = () => {
   return (
-    <div>
-      마이페이지임
-      <Outlet />
-    </div>
+    <>
+      <section className="relative flex h-fit min-h-screen flex-col">
+        <h2>마이페이지임</h2>
+      </section>
+      <MainNavBar />
+    </>
   );
 };
